Inline the import swap into the loader
Five lines don't need their own module and unit test, the e2e case
covers them.
